Various changes; this is the first commit after git nearly killed PokEditor in October 2021

+++ src/com/turtleisaac/pokeditor/framework/narctowl/NarcSubFile.java
@@ -2,8 +2,8 @@ package com.turtleisaac.pokeditor.framework.narctowl;
 
 public interface NarcSubFile
 {
-    int getStartingOffset();
-    int getEndingOffset();
-    int getTrashBytes();
+    long getStartingOffset();
+    long getEndingOffset();
+    long getTrashBytes();
     String getName();
 }
